- use BigDecimal

+++ src/DispenserProvider/Extensions/AmountExtensions.cs
@@ -1,5 +1,6 @@
 ﻿using System.Numerics;
 using DispenserProvider.Services.Handlers.CreateAsset.Models;
+using Nethereum.Util;
 
 namespace DispenserProvider.Extensions;
 
@@ -9,7 +10,7 @@ public static class AmountExtensions
 
     public static string CalculateAmount(this User user, Refund refund) => CalculateAmount(user.WeiAmount, refund.Ratio);
 
-    public static string CalculateAmount(this string weiAmount, decimal ratio) => CalculateAmount(BigInteger.Parse(weiAmount), ratio);
+    public static string CalculateAmount(this string weiAmount, decimal ratio) => CalculateAmount(BigDecimal.Parse(weiAmount), ratio);
 
-    public static string CalculateAmount(this BigInteger weiAmount, decimal ratio) => new BigInteger(ratio * (decimal)weiAmount).ToString();
+    public static string CalculateAmount(this BigDecimal weiAmount, decimal ratio) => (new BigDecimal(ratio) * weiAmount).ToString();
 }
